Destroy and unfreeze
