+++ src/commands/destroy.js
@@ -1,23 +1,16 @@
-const { writeConfig, readConfig } = require('../util/fileUtils');
-const { jadeErr, jadeWarn, jadeLog } = require('../util/logger');
+const { jadeErr, jadeLog } = require('../util/logger');
 const { confirmDestroy } = require('../util/questions');
-const { deleteAllBuckets } = require('../aws/s3/deleteAllBuckets');
-const { deleteIamRole } = require('../aws/iam/deleteIamRole');
-const { deleteLambdaFunction } = require('../aws/lambda/deleteLambdaFunction');
-const {
-  lambdaNames,
-  lambdaRolePolicies,
-  lambdaIamRoleName,
-} = require('../templates/constants');
+const { resetJade } = require('../util/reset/resetJade');
+const { deleteAllApps } = require('../util/deleteApps');
 
 const destroy = async () => {
   try {
     const answer = await confirmDestroy();
     if (!answer) return false;
 
-    deleteAllBuckets();
-    deleteLambdaFunction(lambdaNames);
-    deleteIamRole(lambdaIamRoleName, lambdaRolePolicies);
+    await deleteAllApps();
+    await resetJade();
+    jadeLog('All AWS infrastructure has now been removed.');
   } catch (err) {
     jadeErr(err);
   }
